+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.0 (WIP)
 
 * Add Letsencrypt support using the `greenlock` module
+* Added websocket proxy support
 
 ## 0.1.0